Added to nhmmer the ability to take a fasta file (multiple unaligned sequences) as input. This was done to support phmmer-like ability to search all sequences in one fasta file against all sequences in another. The format is autodetected, such that nhmmer first tries to open a sequence file as an MSA, then if that fails, tries it as unaligned fasta.
Also created a DNA substitution matrix ("DNA1") that approximates the 
posterior emission probabilities computed by nhmmer with the standard
mixture dirichlet combined with a single sequence. This is used by 
default when the nhmmer query is an unaligned fasta file, and can be
forced for MSA inputs of one sequence, using --singlemx.  Modified 
hmmbuild to offer --singlemx as well, and to assign different mx, popen,
and pextend defaults for DNA search.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@860 db3e70e4-12e1-0310-b23f-81a6e3d09ba7