Modifying hmmpgmd to allow a command to be passed a flag (--seqdb_ranges) that will restrict the search to only a subset of the full sequence file. The --seqdb_ranges flag will take a comma-separated list of ranges, e.g. "21..39,121..153". This range-set forces the hmmdwrkrs to search only sequences that (a) match the --seqdb identifier, and (b) fall within one of the given ranges (in the example above, only sequences that fall between the 21st and 39th, or between 121st and 153rd, positions in the target db file.
Also, esl-reformat has been modified to enable creation of a simple hmmpgmd
formatted file (with a single seqdb). This produces both (a) a hmmpgmd file 
with sequence names (headers) going from 1..N for N sequences, and (b) a
file giving the mapping of each sequence id to the original sequence name
and description.  (TO DO: change hmmpgmd to take that mapping file as input, 
and do the name conversion within hmmpgmd)



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@862 db3e70e4-12e1-0310-b23f-81a6e3d09ba7