Manually merged fix to Infernal 1.1 release branch's Easel Makefile.in based on r869 in HMMER's 3.1 release branch. Log for that hmmer revision is:
Fixing a small makefile bug identified by Nils Christian
(nils.christian@uni.lu), sent in an e-mail today. The easel utests,
benchmarks, etc. didn't have ${LIBS} at the end of the cc line, and
should've (it's defined earlier in the file, and was presumably
intended to be used. This seems to be an empty string on the systems
I've looked at, but clearly contained the path to the pthreads library
on Nils' machine, so that 'make check' failed for him.  Adding ${LIBS}
seems to have resolved his problem, and is done here.


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/branches/infernal@870 db3e70e4-12e1-0310-b23f-81a6e3d09ba7