In the H3.1b1 releases, userguide and LICENSE.txt files have been removed, leading to errors in 'make clean' and 'make distclean' when trying to clean them. Added tests to make the requested actions conditional on file/directory existence.
The test uses 'ifneq' instead of 'test -e' because the '+@' 
in QUIET_SUBDIR0 can't be passed to the shell (which is what
happens with 'test -e'). Note that ifneq breaks if indented,
as indentation causes the lines to be treated as part of the 
recipe and sent to the shell. This makes it a smidge less 
beautifully indented than we'd like. I've tried to mitigate
this as much as I can with a little line shuffling.

(this connects to r4489 on the hmmer 3.1 branch)



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/branches/hmmer/3.1@873 db3e70e4-12e1-0310-b23f-81a6e3d09ba7