esl-stranslate bug was submitted by Jaina Mistry.
Two parts:
(1) If any of the 6 frames ended with a stop codon, the
mechanism for counting ORFs counted one too many (it
figured another ORF would follow the stop codon, but
that's not possible if there's no more sequence left).
This led to a sequence with no name, eventually resulting
in a "Segmentation fault: 11" when trying to print that
name.

(2) After fixing that, another problem arose: when each
ORF is created, it is given a name like "Seq_s1_114to134".
The space for that name was allocated prior to using
sprintf to set it. Later, the name was updated to
nucleotide coordinates, again using sprintf. The problem?
Sometimes the new coordinates are one or two characters
longer, but the name field wasn't reallocated. Updated
to using esl_sprintf, which correctly allocates the
space.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/branches/hmmer/3.1@885 db3e70e4-12e1-0310-b23f-81a6e3d09ba7