Some tweaks to esl_buffer's error checking when reading from pipe, found while trying to trace a separate issue (which is that ./esl_buffer_utest -s 325 will occasionally lock up, on my home desktop MacOS/X 10.6.8 with gcc 4.4.2 but not 4.2.1).
Problem was that esl_buffer_utest fails under GDB, but not at command
line, oddly enough; intermittently on desktop, reproducibly on laptop.
It has something do to with feof() and ferror() both returning TRUE
when we call fread() on a popen()'ed stream that's at eof. Therefore,
check for normal EOF first, before checking for error.

This appears to fix the problem but I feel like something else might
be going on here with esl_buffer's handling of input from pipes.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@895 db3e70e4-12e1-0310-b23f-81a6e3d09ba7