Merging changes to esl_msa_ReasonableRF() from infernal's branch into this one. The notes from there were:
~~~

Updating the rmark-create program for the purposes of 
producing a "twilight zone" plot. This requires that I
compute the percent identity between each planted (test)
sequence and some appropriate representative of the query
(train) family. Previously, this was written with plans
to compute only the twilight zone for single sequence
queries, so I just picked one sequence for each family,
and used that as the "train" ... and computed pct_id on
that. When I decided that I wanted to show the benefits 
of profile search, I needed to change this to use a full
"train" alignment, then use the consensus of the alignment
as the single query sequence (and the sequence for which
pct_id is computed). 

In addition to changes in rmark-create, this required 
a modification to esl_msa_ReasonableRF(), a function 
that appears to have been orphaned (not used elsewhere),
but did most of what I needed already. The change 
simply (1) makes the function write the RF line to the
place that the comments say it will (the passed rfline
char*), and (2) gives the option to write the consensus
sequence to the RF line (as Infernal seems to do by
default), rather than a simple run of 'x' and '.'.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/branches/hmmer/3.1@913 db3e70e4-12e1-0310-b23f-81a6e3d09ba7