Simplified how _Validate() and _Compare() routines fail. These are routines that on the one hand, may need to be called from production code so they must return normally w/ a failure code; but on the other hand, when we're debugging, we want them to abort immediately on the line that failed some test.
Had been using two different unsatisfying tricks for this: "killmenow"
flag when DEBUGGING flags were on, and if (killmenow) abort(); or,
ESL_AOF() and ESL_AOG(), "abort or fail" and "abort or goto error
block" which did the same thing. Besides being ugly, disadvantage was
that we had slightly different code paths when in debugging mode.

Simplified by adding esl_fail(), and having the ESL_FAIL macros call
it.  esl_fail() only formats the error message, same as the macros
used to do; but by making it a function, now it's a valid breakpoint
for a debugger. When debugging, set "break esl_fail()", and a Validate
or Compare test will halt on the line that fails a test.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@929 db3e70e4-12e1-0310-b23f-81a6e3d09ba7