one more wrinkle parsing the alignment of a fasta/ssearch ouput with demotic_fasta.pm,
The names of query and target get truncated as they appear in the alingmente part.
That makes it hard sometimes to identify which is query and which is target,
specially if their names share a root. New version solves that ambiguity 
for some examples, but others problems may still appear.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@934 db3e70e4-12e1-0310-b23f-81a6e3d09ba7