Fixed 34 "Dead store/dead assignment" issues detected by clang static analyzer. All were real. 32 were harmless.
2 in esl_msashuffle.c::esl_msashuffle_XQRNA() and
esl_msashuffle_CQRNA() were a real bug. Because of a typo in the
swapping of alignment columns, these shuffling routines were broken,
and would result in corruption of the input alignment. These routines
are not called by HMMER, but others (Eric, Elena?) may want to check.
These routines were not tested by any unit tests (and still aren't) so
we deserve what we get.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@944 db3e70e4-12e1-0310-b23f-81a6e3d09ba7