HMMER bug #h108 fix: nhmmer corrupted by * residues in input seqs.
Two fixes having to do with handling * nonresidue character:

esl_alphabet.c::set_complementarity() - was written for Kp=17 and not
updated when we added * residue to DNA/RNA alphabets, making
Kp=18. However, this problem was likely not the problem in #h108:
result is that [17] is not initialized, and that's ~, not *.

esl_sq.c::esl_sq_ReverseComplement() - similar problem, but for text
mode. * is caught as an invalid character, eslEINVAL return.
esl-sfetch then errors out, refusing to revcomp DNA seq containing *.
This is also not the cause of the nhmmer crashes, but fixed it anyway.



git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/branches/hmmer@946 db3e70e4-12e1-0310-b23f-81a6e3d09ba7