Fixed unit test failures on icc/linux in esl_normal, esl_random.
In esl_normal, the unit test waits for gaussian pdf to drop below
1e-300 to prove that it can cover full double precision range. But
that assumes denormal doubles: at x=37, pdf=2.1e-298, and at x=38,
pdf=1.097e-314. With flush-to-zero on, the test stops a step earlier
than with flush-to-zero off. Turns out that icc turns on flush-to-zero
at all -O levels; gcc does not. The fix is to make the test succeed
whether flush-to-zero is on or not.

In esl_random, the issue is a weird floating point heisenbug that I
don't understand. random roll is [0,1), guaranteed <1, and that seems
to be true. The sampling routine esl_rnd_FChoose depends on a test
roll < cdf[i] / cdf[N-1] eventually succeeding -- and mathematically
it must, for i==N-1, because cdf[N-1]/cdf[N-1] is obviously 1.0. But
for floating point this is somehow not true. It appears that (float)
1.0 < (double) 1.0 at least in some cases (!?). It's a heisenbug,
because inserting a printf() to print the values being tested makes
the bug go away. The fix is to cast the cdf[]'s: roll < (double)
cdf[i] / (double) cdf[i]. If someone can explain why that works, I'd
appreciate it.