Update to the ReadBlock functions, impacting longtarget (nhmmer) blocks. Previously, blocks could differ almost 2-fold in length. Now the slop is only 5%.
Previously, the function worked like this (for longtarget/DNA seqs):
- goal: get a bunch of sequence with total length roughly max_residues
- Since we're dealing with DNA, we're happy to extract windows of sequence
  from a (much) longer sequence. But we also might be dealing with short
  sequences, so we need to be willing to grab a bunch of them
- Request a window of up to max_residues, add it to the block list
- If the received window is shorter than than max_residues, ask for
  another window of up to max_residues.
- Continue requesting windows (each time asking for up to max_residues)
  until the total length is > max_residues.

This runs the risk of producing a set of sequences with total length
almost 2*max_residues (if the limit is almost reached by one window,
then extended by another max_residues on the next window). This is fixed
by changing how we request extension. Trivially, this could just ask
for a window of at most (max_residues-total_size_so_far) ... but there
are cases where this ends up requesting a tiny prefix of a very long
sequence. To avoid that, I've allowed a bit of slop, so that the next
window request will be at least 5% of max_residues:
    max( (max_residues - size_so_far) , (0.05*max_residues) )
This ensures that all windows are reasonably long, and that any two
blocks will be roughly the same length.

This fix will allow me to clean up an issue with wasteful memory
allocation in the FM-index construction phase of makehmmerdb.