Fixed formatting bug, all tests passing
There was a formatting error in the esl-translate test caused
by vsnprintf call to store the sequence description. For some
reason, x86 can handle the %d format but ARM prefers %lld for
that. In the future, we should watch out for these errors. At
this point all tests are passing and the code is ready to be
optimized and merged into the arm branch, safe for possibly a
few configuration changes and changes to the build targets.