Removes easel's "augmentation" concept; adds safety from gnu autoheader.
Since its inception Easel had an overly cute concept called
"augmentation", whereby modules had a base configuration that could be
stacked with additional modules to add more superpowers. The intent
was to allow easier disentanglement of an Easel .c file from the rest
of the library, if someone wanted to use a small part of our code, not
all of it. I was especially thinking of teaching the WashU Bio5495
course, which was taught in C. However, the #ifdef's required for
augmentation add complexity, and create a combinatorial mess of code
paths that increases testing difficulty; and now that I teach in
Python, the cost outweighs the benefits. After I added
esl_msa_Sample() as an augmentation to msa a few days ago, I made a
note to get rid of augmentation altogether, and here we go.

Also spurred by a bug report from Michael Crusoe of Debian. Debian
runs 'autoreconf' routinely, which runs 'autoheader', and our code was
not built to withstand the evil ravages of gnu autoheader. There are
two issues. 1) autoheader requires all AC_DEFINE() macros to use their
full 3-arg form. On our configure.ac, it errors out. 2) autoheader
creates a .h.in file based on the first AC_CONFIG_HEADERS() that it
sees. On our configure.ac, it overwrites our esl_config.in.

- Removed --enable-modular arg from configure
- Removed all augmentation #ifdef's, and documentation about augs
- Deleted devkit/esl-dependencies, a tool for tracking required
  vs. augmented header file inclusion
- Add decoy_config.h.in and AC_CONFIG_HEADERS([decoy.config.h]) to
  distract autoheader from overwriting our code
- Moved esl_msa_CreateFromString() to esl_msafile, to avoid a
  circular dependency msa <-> msafile
- Made sure esl_config.h is included first in all Easel files