Adds esl_sse_{right,left}shiftz_float() functions.
When doing a shift that needs to shift a 0.0 value on, a minor
optimization becomes possible, because IEEE754 representation of zero
is all 0 bits: so we don't have to OR against any neginfmask, we can
just to a vector bit shift instruction and we're done. Call this
variant "shiftz".

We have to preserve the old esl_sse_{right,left}shift_ps() functions
because Infernal uses them extensively.