Fixes bug in ESL_AVX512 configure macro that could overwrite CFLAGS.
This is the same one-line fix as in commit 745a9a1 on the release-0.44
branch. I'd meant to apply it to `develop`, where Nick is working.
Don't want to merge release-0.44's other changes back yet, so decided
it was easier to just make the same change as a separate commit. Watch
out for this on merge, but it should be fine.