Adds `easel` miniapp and its first subcommand, `easel downsample`.
I'm going to start switching the Easel miniapps slowly over to the new
subcommand style, `easel <subcommand>` instead of `esl-<subcommand>`.
HMMER4 will also use this style with `hmmer <subcommand>`.

`easel downsample` is the first such miniapp. It adapts esl-selectn,
and extends it to be able to sample sequences, not just lines.