Adds autodoc.py: extract function documentation in Markdown
I'd previously used `autodoc`, a Perl script, to extract my structured
function header information to LaTeX. I don't think this has been
terribly effective over the years. I want to try switching to simpler
Markdown and on-line documentation instead - and indeed, I promised it
in our current NIH grant. This is a first draft of a new Python script
that takes a .c file as input and generates a Markdown .md file with
function documentation.