Adds ESL_STR(), a preprocessor trick to stringify defined constants.
ESL_STR(), with its partner ESL_XSTR(), converts a #define'd constant
(as a number) to a string constant. The string constant can then be
used as a default setting in a getopts OPTIONS array, for example.
This helps us #define default parameters in one place. The macro pair
uses a trick with the C preprocessor's stringify operator.