Adds Easel technology tree figure.
easel_techtree.ai shows a high-level overview of all Easel modules,
organized into groups according to their dependencies.

I'm aiming for this to be a DAG, and it almost is - but currently the
sequence file and alignment file groups are a tangled mess that will
someday need to be sorted out.

Also, the gencode module has a dependency on getopts that the figure
doesn't show. Added a TO DO note in esl_gencode.c that this dependency
should be removed; optional configuration params should use a _CFG
structure the way esl_msaweight does it, rather than dictating exact
commandline options to a program -- a mistake that I'm walking back
here and elsewhere (in HMMER's pipeline, for example).

xref 2019/0415-easel-tech-tree-v3. I've had the figure for a while,
but this is the first time I've put it into the repo.