Makes fasta format parsing more robust against .afa files
Under some circumstances -- via esl_sqio_SetDigital() -- the fasta
parser could get set in a state that allowed gap characters as legal
input, because the abc_inmap did. If the user provides an aligned
fasta file with a .fa suffix, inadvertently claiming that it's a FASTA
file, the parser would then read gaps as valid symbols, and this could
confuse a program like hmmsearch that needs unaligned seq input.

For all unaligned parsers, set the inmap for '-' to eslDSQ_ILLEGAL.

Also, tightens up the suffix rule for detecting "*.fa" so it detects
only "*.fa", not just a .fa prefix on the suffix. (Same for .gb.)

This is the fix for HMMER issue 153.