Fixes a esl_json_utest failure due to icc's handling of denormals
esl_json_utest's utest_read_float() included a test for proper reading
of a denormalized float. However, denormals may be flushed to zero
depending on the compiler, and indeed icc's default behavior (with
optimization on) is to set the DAZ and FTZ cpu flags, so
esl_json_utest failed when compiled with icc (but not gcc, clang).

Removed the test. utest_read_float() is already compulsively thorough
enough.