Documents why the vector ISA autoconf tests are compile tests, not run tests.
Responding to a pull request that wanted to change AC_LINK* to AC_RUN*
tests. Use of AC_LINK is deliberate; we're testing whether the
_compiler_ supports certain ISAs, not whether the processor does. Our
code uses runtime dispatching to detect what ISAs the processor
supports.  For example, it is not a bug if a build environment detects
AVX2 support (in the compiler) while not supporting AVX2 on its
processor.