This commit contains the Easel parts of the changes to fix the makehmmerdb memory use bug.  There are two aspects to this change:
1) esl_sq_BlockReallocSequences() has been added to esl_sq.c
and esl_sq.h.  This function reallocs all of the variable-sized elements of
the block's sequence list back to their default sizes.  Makehmmerdb and nhmmer
should call this function when the size of their ESL_SQ_BLOCK objects
exceed some threshold to counteract the "slow growth" memory use bug that
occurs when long sequences are fetched into many of the block's sequence
objects because the same block object(s) are re-used many times in processing
a large sequence database.

2) sqascii_ReadBlock() has been changed so that, instead of reading sequences
directly into the block's sequence list using sqascii_ReadWindow(), it fetches
sequences into a temporary sequence object and then copies them into the block,
except for the first sequence fetched, which is read directly into the block
to reduce the performance impact of the double copy when reading long
sequences that span multiple blocks.

This change addresses the "fast growth" problem where a single call to
ReadBlock could result in the block occupying multiple terabytes of memory.
This occurred because ReadWindow() reads max_residues bytes of the sequence
file into its target sequence object, determines where the sequence being
fetched actually ends, and sets the length field of the object appropriately.
When ReadBlock() needed to fetch many short sequences to meet its residue
quota, this led to the block's sequence list containing many sequences that
occupied max_residues bytes of memory, exploding the size of the block.

Reading sequences into a temporary sequence object eliminates this problem
because esl_sq_Copy() only copies the bytes that represent the residues in
a sequence, regardless of how large the source sequence's buffers are.