Adds a note to esl_msaweight_PB() documentation.
esl_msaweight_PB() defaults to using RF consensus column annotation if
present, else it defines its own consensus cols. So the same MSA with
versus without RF annotation may get different consensus cols. Easel
views this as a desirable feature, because PB weighting is faster if
RF annotation is provided, and it views MSAs with vs. without RF
annotation as different inputs. However, in HMMER and Infernal we
don't want model parameterization to depend on presence/absence of RF
annotation; in default hmmbuild/cmbuild, we want to ignore any RF
annotation, and with --hand, we want to use it. (And in jackhmmer
iterations, which are always --hand, we want to use it.) HMMER and
Infernal have been modified to use a custom esl_msaweight_PB_adv()
configuration that sets ignore_rf to FALSE when --hand is set, else
to TRUE. [see EddyRivasLab/hmmer#180 for discussion with Eric]

This commit adds a note to the documentation of esl_msaweight_PB(),
memorializing this discussion.