fixes iss #56: memory fault with zero-length sequence descriptions, etc.
Several of the esl_msa_Set*() functions can take either a
NUL-terminated string or a non-terminated buffer <buf>, <n> as
input. When the input buf is non-terminated, length <n> is required;
when it's a NUL-terminated string, the caller can either pass its
length if known, or n=-1.

n=0 is a valid (albeit unusual) buffer length, but these functions
were erroneously testing for n>0 not n>=0 to differentiate between
buffers (or strings) to handle with esl_memstrdup() versus strings to
handle with esl_strdup().  esl_strdup() tries to call strlen(), which
is a memory problem on an unterminated buffer.