esl_{FD}Compare(), esl_{FD}CompareNew() renamed; esl_{FD}CompareAbs() removed.
The old esl_FCompare() and esl_DCompare() functions for comparing
floating point values for equality have been marked as deprecated for
some time, and esl_FCompareNew() and esl_DCompareNew() have been
recommended instead.  Whereas the old Compare functions use only a
relative tolerance, the CompareNew() functions use a combination of
relative and absolute tolerance. This seems to be a more careful if
not common practice in numerical computing.

This commit swaps function names:
   esl_FCompare()    => esl_FCompare_old()
   esl_FCompareNew() => esl_FCompare()
   esl_FCompareAbs() => (deleted)

And likewise for their D (double) versions.

This commit will break any code that uses these functions. The easiest
way to upgrade:

1. Replace all occurrences of esl_FCompare() with esl_FCompare_old().

2. Replace all occurrences of "esl_FCompareAbs()" with "esl_FCompare()" and add
   a relative tolerance of 0.0 as the next-to-last arg, e.g.
     esl_FCompareAbs(x0, x, abs_tol) => esl_FCompare(x0, x, 0.0, abs_tol)

3. Replace all occurrences of esl_FCompareNew() with esl_FCompare().

Results will be unchanged if you update code this way.
esl_FCompareAbs() can be safely replaced by esl_FCompare() with a rtol
of 0.0; the calculation is identical.

The reason to rename to esl_FCompare_old() and keep the old function
around is that an rtol-only comparison (with abstol=0.0) with the new
esl_FCompare() is not the same calculation as the old rtol-only
comparison with what's now esl_FCompare_old().

The old version calculates relative difference as fabs(a-b) /
(fabs(a+b)/2) ... i.e.  the denominator is the average of the values
being compared. Order of the a,b values doesn't matter, and it doesn't
matter too much if one of them is 0.0.

The new version calculates relative difference as fabs(x0-x) /
fabs(x0) ...  i.e. one of the two values is used as the reference
"true" value.  The order of the arguments matters, and if the
reference value x0 is 0, the only way to pass an rtol-only comparison
is if x0=x exactly.  You're supposed to use a combination of rtol and
abstol, so comparisons for x0~0 depend on abstol.

Despite what it looks like, it doesn't ever divide by zero. The test
actually is:
   if (fabs(x0 - x) <= r_tol * fabs(x0) + a_tol) return eslOK;

This difference means that you can't just swap in the new rtol+abstol
FCompare() without risking breaking something. Thus I'm keeping the
old versions around for now. We can replace calls to them over time.