Adds esl_resize(): idiomatic resize-by-doubling + redline allocation rule.
Many structures are resizeable. I often follow an idiom of
reallocating by doubling; realloc() is expensive and exponential
reallocation strategies are more efficient than linear ones.  I also
often impose a "redline", to cap or at least minimize exceptionally
large reallocations. A redline is especially useful in cases where an
object is being reused for an iterated set of tasks with a long-tailed
distribution (protein sequence lengths, for example); an object can be
briefly overallocated to an exceptional size, then pulled back down by
its _Reuse() call.

This commit introduces esl_resize() to implement the logic of this
idiom in a single place, and to implement it in a way that protects
the doubling strategy from integer overflow.

Also adds a utest_resize() unit test, and updates the easel.c test driver.