Fixes an intermittent esl_gamma unit test failure.
The esl_gamma FitComplete unit test can stochastically fail, but it's
supposed to reinit the RNG with a fixed seed to avoid that. It was
sampling lambda and tau params before that reinit.

Additionally, even stochastic failure is too frequent. Looked into it;
when tau is small (<0.2 or so), the distribution is very peaked at its
base at mu, and the fit is insensitive to relatively large changes in
tau or lambda. To avoid this issue, I changed the sampled range for
tau and lambda in the FitComplete test from (0.,10.) to (0.5,10.).

Improved documentation: for example, that gam_nll() is the
one-parameter form (only dependent on tau, not lambda), as described
in esl_gamma.md, because this initially looked like an error to me,
and it's not.

[SRE:H13/23]