Silences compiler warning in esl_threads.c::esl_threads_CPUCount() on Linux/gcc
Linux/gcc has deprecated sysctl() and issues an annoying warning about
it.  We've been using sysctl() as a fallback to obtain the CPU core
count, if a machine lacks sysconf(). (My notes indicate that some
BSD-based platforms did.) sysconf() is a POSIX.1 standard, so it's
reasonable to rely on it being present. I removed the sysctl() code
from esl_threads_CPUCount() and silenced the warning.