Revises the esl_vectorops::stable_sums unit test
This unit test "failed" on icc with optimizations on, because a
numerical result that was expected to fail (a naive summation)
actually came out correctly.  By default, icc has unsafe fast-math
optimizations on (-fp-model=fast), and it's allowed to rearrange
calculations in ways that affect floating point results. gcc, in
contrast, has fast-math optimizations off.

I revised the unit test to allow the naive summation to either succeed
or fail; but if it fails, Kahan and Welford compensated summation
algorithms must still succeed.