Fixes clang compiler warnings in esl-ssdraw.c's use of SSDRAWINFINITY
SSDRAWINFINITY is defined as a large integer (987654321), but it is
used everywhere as a float. clang warns about implicit conversion from
int to float, including calls to esl_FCompare_old() which appear to
require a comparison of inf to inf to return eslOK. SSDRAWINFINITY is
too large to be stored exactly in a float; the comparisons have
probably been working because either it's always implicitly converted
to the same float, despite not being exactly 987654321.0, or because
it's within the roundoff error range for esl_FCompare_old().

It appears that I can just replace SSDRAWINFINITY with eslINFINITY, so
I did.  (Eric will want to check and verify.)