Fixes thread race condition in esl_threads.
Detected in esl_threads_example by both helgrind and tsan. The problem
is that WaitForFinish() was decrementing threadCount one at a time as
threads finished, but newly started workers in Started() need
threadCount to be the total number of assigned workers, not just the
current number of running workers, because they access the array of
work units. This meant that if one thread finished before another one
started, there was a problem. This wouldn't normally happen (threads
have work to do, and they start faster than anyone's work should take)
but still.

Also fixes a compiler warning about a misleading indentation in esl_alloc.

[SRE H14/104-105]