Fixes multithreading issue in esl_stack, not robust to spurious wakeups
pthread_cond_wait() may return upon a "spurious wakeup", not just
because someone sent the condition. So whenever we're having a thread
send a condition signal to another, we ordinarily have a trio of a
mutex, a condition, and a (often boolean) variable, and the pattern
goes like:

Sender:

   get mutex
   change variable
   send condition
   release mutex

Receiver(s):

   get mutex
   while (variable not changed)
     pthread_cond_wait() for condition
   release mutex

esl_stack poppers were only using an if, not a while, so they weren't
correct against spurious wakeup. Where they had:

 if (s->do_cond && ! s->n && pthread_cond_wait(s->cond, s->mutex) != 0) ESL_EXCEPTION(eslESYS, "pthread_cond_wait() failure");

I changed these to:

 while (s->do_cond && s->n == 0) { if (pthread_cond_wait(ns->cond, ns->mutex) != 0) ESL_EXCEPTION(eslESYS, "pthread_cond_wait() failure");  }

The master pusher(s) wake the poppers for two reasons. One is that
they've pushed new work; now s->n > 0. The other is that there's no
more work. The cond_wait while loop needs to be able to distinguish
this from a spurious wakeup with a temporarily empty stack. Thus a
master calls esl_stack_ReleaseCond() to signal that there's no more
work. This function sets do_cond = FALSE then broadcasts the condition
to all sleeping poppers.

For this to work with the while loop properly, ReleaseCond() has to
set do_cond = FALSE *before* broadcasting the condition, which it
wasn't doing. Fixed this too.