Adds esl_vec_?Shuffle64() routines to guarantee uniform sampling, including large vectors.
In #9edb096, I changed n from int to int64_t, after Eric reported a
bug when v is a large vector of >2B elements, and the implicit cast of
the int n argument caused an overflow error. However, I also noted
that this fix worsened a problem with achieving uniformly distributed
shuffles. This is because esl_rnd_Roll(rng, n) takes a uniformly
distributed 32-bit esl_random() sample on [0,1) and simply multiplies
it by n. This only produces a uniform sample on [0,n) if n is a power
of 2, and if n <= 2^32.

We've known about the issue with n not being a power of 2. It was
noted in esl_rnd_Roll(), along with a note that esl_rand64_Roll() is a
correct implementation for any n. For n > 2^32, the problem is in
using the ESL_RANDOMNESS generator at all, because it's a 32-bit
generator. We could concatenate two 32-bit random variates, I
suppose. But the better solution may be to have a set of alternative
large-vector shuffling functions that use the ESL_RAND64 64-bit
generator.

Changes:

Added esl_vec_{DFIL}Shuffle64() functions, which take an ESL_RAND64
64-bit RNG as their first arg, and which guarantee uniformly
distributed shuffles even for large vectors of >2B elements.

Moved esl_random::esl_rnd_mix3() to easel:esl_mix3(). Now esl_rand64
is independent of esl_random; removed #include "esl_random.h" from
it. Used cscope to verify that esl_rnd_mix3() is only called in
esl_rand64, esl_random in Easel|HMMER|Infernal.

Replaced esl_rnd_Roll() macro with an esl_rnd_Roll() function that
guarantees uniformity of its sampled distribution for any n, not just
when n is a power of 2, following esl_rand64_Roll()'s
implementation.

In techtree.ai documentation of Easel dependency graph, moved
esl_rand64 from "numerical methods" to "base", and move esl_vectorops
from "base" to "algorithms & data structures". esl_vectorops depends
on esl_random already, though I have it classed as a "base" Easel
module, and now I'm making it dependent on esl_rand64 too because of
the Shuffle64 functions.

esl_recorder_utest failed in utest_block() because it can call
esl_rnd_Roll() with n=0, when minalloc > nalloc. New esl_rnd_Roll()
enforces n>=1. Added an if to avoid this case.