Fixes baffling esl_rand64_utest failure, specific to icx -O3 -fp-model=precise
Our code assumes IEEE754-compliant floating point operations. One
important example is that we use -eslINFINITY as a valid log
probability value when a probability is 0, so we need -infty == -infty
to evaluate to TRUE. Most compilers are IEEE754-compliant by default,
including gcc and clang, but Intel compilers default to a non-IEEE754
compiliant mode (-fp-mode=fast). A default Intel compilation generates
a slew of compiler warnings on our code because it sees us doing
-inf==-inf comparisons that it thinks "always evaluate to FALSE". With
the Intel compiler, a user needs to use n option to turn IEEE754
compliance on, such as `-fp-model=precise` or (probably more safely)
`-fp-model=strict`; for example, `./configure CC=icx CFLAGS="-O3
-fp-model=strict`.

With `-O3 -fp-model=precise`, another issue arises: esl_rand64_utest
fails. The failure is really baffling; I couldn't figure out what was
causing it. Whatever the problem is, it's occurring in code that the
Intel compiler has optimized out with -O3. I can't trace the problem
in the debugger. And when I use printf()'s, the problem disappears.
Xref [2023/0803-h3-icx-fpmodel] for gory details. I'm going to blame
the Intel compiler/linker; I don't believe it's a problem in our code.

I did stumble across a workaround, one that only makes the problem
more baffling: it goes away if I #if out the entire main code blocks
so they don't recompile when we compile the BENCHMARK, TESTDRIVE, and
EXAMPLE drivers in our idiomatic use of eslRAND64_BENCHMARK, etc
flags. Unclear why this would do anything - more evidence that the
problem is in the compiler/linker/optimizer? - but since it works, I'm
using it.