Documents why we have a ptrdiff_t cast in ESL_ALLOC(p, (ptrdiff_t) sizeof(foo)*n) calls
`malloc(sizeof(foo) * n)` can trigger annoying warnings on some
versions of gcc that have a `-Walloc-size-larger-than=PTRDIFF_MAX`
option turned on, because of aggressive checking of a problem that can
arise in unsigned/signed arithmetic if n is a negative signed
integer.

Added notes at ESL_ALLOC() in easel.h and in the code style document
codestyle.md to recommend a cast of sizeof() to an appropriate signed
int type, in an idiom of ESL_ALLOC(p, (ptrdiff_t) sizeof(foo)*n).

The problem is explained more in HMMER commit #b352c6bd, and
in excruciating detail in (xref SRE:2023/0804-h3-gcc-malloc-warning).