Fixes an overzealous asan leak warning in esl-mask.itest.pl
The itest runs `esl-mask -h`, which prints help and exits, and asan
sees a leak there. Well, fine, all our miniapps leak memory with `-h`,
by design; since they exit immediately with the help message, they
just don't bother to free the trivial initial allocations at program
start, typically ESL_GETOPTS. But asan isn't complaining about
ESL_GETOPTS being leaked (though it is), or it would complain on *all*
our itests.  It complains about `esl-mask` specifically - and only
when compiled with optimization on, not with `--enable-debugging`.
The difference with `esl-mask` is apparently that it happens to also
allocate an ESL_SQ before it checks for the -h command line flag. I
moved this allocation to after the `-h` option is handled.  Problem
solved.