Fixes issue with truncated option lists in esl_getopts error messages.
This is the fix for HMMER issue 317, from Nick.

If incompatible options are used, and if the list of incompatible
options is long, the error message formatted by esl_getopts may be
truncated. (Similarly, if a required option is missing, with a long
list of required options, though that would probably be rarer.)

The limitation here is that ESL_FAIL error messages have a fixed max
length of eslERRBUFSIZE-1, i.e. 127. We have to be stingy in how we
include any long strings (file paths; option lists) in formatted error
messages.

- Change error messages for required/incompat opts to print just the
  one conflicting required/incompat option name, rather than the entire
  list (which could be quite long).

- Document that max option name length is 24, in esl_getopts.md

- Add optname length check in esl_getopts_Create()