Adds --enable-msan to ./configure, to enable memory sanitizer
A third sanitizer option, along with the thread and address
sanitizers.  Unlike the thread sanitizer (--enable-tsan) and the
address sanitizer (--enable-asan), the memory sanitizer is only
supported on clang and Linux, not gcc or MacOS.

CC has to be set to clang too to use this. I didn't want to change the
order that we do things in the configure script, and we select and
test the compiler before setting sanitizer flags. So to use MSAN you
would do:

   ./configure CC=clang --enable-msan

I'm not checking for Linux and clang when --enable-msan is set;
compilation will just fail instead.