Adds checks for non-ASCII chars in Stockholm #=GC, #=GR annotation
Encountered a case where an RNA stem was annotated with P1, P1’ in
this case).  That ’ in P1’ is not ', it's UTF-8's U-2019 RIGHT SINGLE
QUOTATION MARK. UTF-8 multibyte encodings are madness in an alignment
where we depend on one character per column. The Easel failure message
was unhelpful; the parser complained that the #=GC annotation was the
wrong length, even though in a UTF-8 capable editor (emacs), it looked
fine.  Added a check in stockholm_parse_{gc|gr} to detect non-ASCII
chars and issue a useful failure message.

The check uses isascii(), which is probably the right way to check a
char without needing to know if the platform defaults to signed or
unsigned char.

Deliberately did not add isascii() checks more generally for Stockholm
annotations. I'm sure Pfam/Rfam have picked up UTF-8 encodings in a
bunch of places, especially PubMed references. The Easel parser will
faithfully propagate them. The problem is only in aligned annotation
where we expect one byte per column (#=GC, #=GR, and the sequence
itself).  The sequence is already validated, at least for digital
mode input.