Changes to allow empty Stockholm MSAs (nseq = 0)
A long-standing one from Tom's chamber of horrors.  135 MSAs in
current Pfam full MSAs (v37.3) have 0 sequences in them. This is dumb
but we should deal with it gracefully rather than failing out, so we
can still index and retrieve MSAs from Pfam full.

This requires a change in how esl_msafile_Read() works, which we use
all over the place, so it's a little scawy to make any changes, but
here we go. My strategy is to handle an empty Stockholm alignment by
returning an error code (eslENODATA), while returning a valid ESL_MSA
structure with header metadata in it, similar to how ESL_SQ can be an
"info" structure with no sequence data. This way, existing calls to
esl_msafile_Read() should continue to see an empty MSA as an error
(because of the non-eslOK return status), and callers that want to
allow for empty MSAs can be rewritten to look for the new eslENODATA
return.

Changes esl_msafile_stockholm_Read() to return eslENODATA and a valid
ESL_MSA with header metadata and nseq=0, when a Stockholm MSA is
empty.

Changes esl_msafile_Read() to do the same. Only Stockholm format
allows for empty alignments, and only the Stockholm reader can return
eslENODATA.

Changes esl_msafile_ReadFailure() to generate an appropriate error
message in the eslENODATA case -- similar to what we did before in the
nseq=0 case from within esl_msafile_stockholm_Read().

Changes esl_msa_Validate() to allow ESL_MSA to have nseq=0.

Updates unit tests in esl_msa_stockholm.

easel aindex|afetch|msastat modified to allow empty MSAs.